--- app/src/main/java/com/pedrenrique/githubapp/core/platform/ModelHolder.kt
@@ -1,5 +1,0 @@
-package com.pedrenrique.githubapp.core.platform
-
-abstract class ModelHolder {
-    abstract fun type(typesFactory: TypesFactory): Int
-}
