+++ app/src/main/java/com/pedrenrique/githubapp/features/common/adapter/ModelHolder.kt
@@ -1,0 +1,7 @@
+package com.pedrenrique.githubapp.features.common.adapter
+
+import com.pedrenrique.githubapp.features.common.adapter.factory.TypesFactory
+
+abstract class ModelHolder {
+    abstract fun type(typesFactory: TypesFactory): Int
+}
