+++ app/src/main/java/com/pedrenrique/githubapp/features/common/adapter/viewholder/EmptyViewHolder.kt
@@ -1,11 +1,10 @@
-package com.pedrenrique.githubapp.features.common.viewholder
+package com.pedrenrique.githubapp.features.common.adapter.viewholder
 
 import android.view.View
 import com.pedrenrique.githubapp.core.data.Failure
-import com.pedrenrique.githubapp.core.platform.BaseViewHolder
-import com.pedrenrique.githubapp.core.platform.TypesFactory
-import com.pedrenrique.githubapp.features.common.model.EmptyModelHolder
-import com.pedrenrique.githubapp.features.common.model.RepositoryModelHolder
+import com.pedrenrique.githubapp.features.common.adapter.BaseViewHolder
+import com.pedrenrique.githubapp.features.common.adapter.factory.TypesFactory
+import com.pedrenrique.githubapp.features.common.adapter.model.EmptyModelHolder
 import kotlinx.android.synthetic.main.item_pull_requests_empty.view.*
 
 class EmptyViewHolder(view: View) : BaseViewHolder<EmptyModelHolder>(view) {
