+++ app/src/main/java/com/pedrenrique/githubapp/features/common/adapter/viewholder/ErrorItemViewHolder.kt
@@ -1,9 +1,9 @@
-package com.pedrenrique.githubapp.features.common.viewholder
+package com.pedrenrique.githubapp.features.common.adapter.viewholder
 
 import android.view.View
-import com.pedrenrique.githubapp.core.platform.BaseViewHolder
-import com.pedrenrique.githubapp.core.platform.TypesFactory
-import com.pedrenrique.githubapp.features.common.model.ErrorItemModelHolder
+import com.pedrenrique.githubapp.features.common.adapter.BaseViewHolder
+import com.pedrenrique.githubapp.features.common.adapter.factory.TypesFactory
+import com.pedrenrique.githubapp.features.common.adapter.model.ErrorItemModelHolder
 import kotlinx.android.synthetic.main.item_error.view.*
 
 class ErrorItemViewHolder(view: View) : BaseViewHolder<ErrorItemModelHolder>(view) {
