+++ app/src/main/java/com/pedrenrique/githubapp/features/common/base/BaseListFragment.kt
@@ -1,14 +1,14 @@
-package com.pedrenrique.githubapp.features.common
+package com.pedrenrique.githubapp.features.common.base
 
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.pedrenrique.githubapp.core.ext.defaultFriendlyMsg
-import com.pedrenrique.githubapp.core.platform.Adapter
-import com.pedrenrique.githubapp.core.platform.EndlessRecyclerViewScrollListener
-import com.pedrenrique.githubapp.features.common.model.ErrorItemModelHolder
+import com.pedrenrique.githubapp.core.ext.defaultFriendlyTitle
+import com.pedrenrique.githubapp.features.common.adapter.Adapter
+import com.pedrenrique.githubapp.features.common.listeners.EndlessRecyclerViewScrollListener
+import com.pedrenrique.githubapp.features.common.adapter.model.ErrorItemModelHolder
 
 abstract class BaseListFragment : Fragment() {
 
@@ -20,7 +20,10 @@ abstract class BaseListFragment : Fragment() {
     protected fun RecyclerView.setup(adapter: Adapter, loadMore: () -> Unit) {
         layoutManager = LinearLayoutManager(context)
         endlessRecyclerViewScrollListener =
-            EndlessRecyclerViewScrollListener(linearLayoutManager, loadMore)
+            EndlessRecyclerViewScrollListener(
+                linearLayoutManager,
+                loadMore
+            )
 
         val decoration = DividerItemDecoration(context, linearLayoutManager.orientation)
         addItemDecoration(decoration)
@@ -34,7 +37,7 @@ abstract class BaseListFragment : Fragment() {
         val error =
             (lastOrNull { it is ErrorItemModelHolder } as? ErrorItemModelHolder)?.error?.error
         error?.run {
-            Toast.makeText(context, defaultFriendlyMsg, Toast.LENGTH_LONG).show()
+            Toast.makeText(context, defaultFriendlyTitle, Toast.LENGTH_LONG).show()
         }
     }
 }
