--- app/src/main/java/com/pedrenrique/githubapp/features/common/model/LoadingItemModelHolder.kt
@@ -1,10 +1,0 @@
-package com.pedrenrique.githubapp.features.common.model
-
-import com.pedrenrique.githubapp.core.data.Loading
-import com.pedrenrique.githubapp.core.platform.ModelHolder
-import com.pedrenrique.githubapp.core.platform.TypesFactory
-
-class LoadingItemModelHolder : ModelHolder() {
-    override fun type(typesFactory: TypesFactory) =
-        typesFactory.type(Loading.Item)
-}
