+++ app/src/main/java/com/pedrenrique/githubapp/features/repositories/RepositoriesFragment.kt
@@ -4,21 +4,16 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.navigation.Navigation
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.pedrenrique.githubapp.R
 import com.pedrenrique.githubapp.core.data.Failure
 import com.pedrenrique.githubapp.core.data.Repository
 import com.pedrenrique.githubapp.core.ext.supportActionBar
 import com.pedrenrique.githubapp.core.ext.supportActivity
-import com.pedrenrique.githubapp.core.platform.Adapter
-import com.pedrenrique.githubapp.core.platform.EndlessRecyclerViewScrollListener
-import com.pedrenrique.githubapp.features.common.BaseListFragment
-import com.pedrenrique.githubapp.features.common.TypesFactoryAdapter
+import com.pedrenrique.githubapp.features.common.adapter.Adapter
+import com.pedrenrique.githubapp.features.common.base.BaseListFragment
+import com.pedrenrique.githubapp.features.common.adapter.factory.TypesFactoryAdapter
 import kotlinx.android.synthetic.main.fragment_repositories.*
 import org.koin.androidx.scope.currentScope
 import org.koin.androidx.viewmodel.ext.android.viewModel
